ignore client shutdown exceptions when sending responses
patch from Ulrik Miaelsson

If the user clicks the stop button, or click another link before
the previous has finished loading, or something similar an IOError
exception will be raised which results in the admin being sent an
email.

This can understandably be pretty annoying if your users are
doing that on a regular basis. So we'll trap that exception
and ignore it.


git-svn-id: https://svn.roundup-tracker.org/svnroot/roundup/trunk@3942 57a73879-2fb5-44c3-a270-3262357dd7e2