clean up query display of "Private to you" items
Previously the handling of queries in the edit menu was a little wonky. If
you change a query to not be private to you then you lose the ability to
ever change it back to being private.

Previously the queries were displayed in three chunks: First all retired
queries. Then all queries private to the current user (this batch has the
dropdown for toggling private state). Then all non-private queries.

Now: all queries we created that are retired, all queries that we created
that are not retired, then all non-private queries not created by the
current user.

this should fix sf bug 1481394


git-svn-id: https://svn.roundup-tracker.org/svnroot/roundup/trunk@3959 57a73879-2fb5-44c3-a270-3262357dd7e2