handle bad cookies
Roundup's cgi gets all cookies from all apps on the same server. some
apps aren't as well behaved as roundup and generate cookies that do
not follow the spec perfectly. This causes python's Cookie module to
throw an exception when trying to parse them. In the spirit of being
liberal in what we accept, we create a LiberalCookie class that can
handle those out-of-spec cookies and allow roundup to continue onward.

sf bug 1691708
code from S. Woodside


git-svn-id: https://svn.roundup-tracker.org/svnroot/roundup/trunk@3962 57a73879-2fb5-44c3-a270-3262357dd7e2