assorted pyflakes fixes
I (re-)discovered pyflakes today and ran it over to roundup
code to see what it had to say. It reported a ton of "imported
module foo but you're not actually using it" warnings and a couple
of actual errors that this patch fixes.

The code was wrong but it was mostly stuff that never gets
executed...like hyperdb.Choice.


git-svn-id: https://svn.roundup-tracker.org/svnroot/roundup/trunk@3970 57a73879-2fb5-44c3-a270-3262357dd7e2